+++ R/edit.R
@@ -3,6 +3,8 @@
 #' Opens a file for editing in RStudio, if that is the active environment, or
 #' via [utils::file.edit()] otherwise. If the file does not exist, it is
 #' created. If the parent directory does not exist, it is also created.
+#' `edit_template()` specifically opens templates in `inst/templates` for use
+#' with [use_template()].
 #'
 #' @param path Path to target file.
 #' @param open Whether to open the file for interactive editing.
@@ -33,6 +35,43 @@ edit_file <- function(path, open = rlang::is_interactive()) {
     utils::file.edit(path)
   }
   invisible(path)
+}
+
+#' @param template The target template file. If not specified, existing template
+#'  files are offered for interactive selection.
+#' @export
+#' @rdname edit_file
+edit_template <- function(template = NULL, open = rlang::is_interactive()) {
+  check_is_package("edit_template()")
+
+  if (is.null(template)) {
+    ui_info("No template specified... checking {ui_path('inst/templates')}")
+    template <- choose_template()
+  }
+
+  if (is_empty(template)) {
+    return(invisible())
+  }
+
+  path <- proj_path("inst", "templates", template)
+  edit_file(path, open)
+}
+
+choose_template <- function() {
+  if (!is_interactive()) {
+    return(character())
+  }
+  templates <- path_file(dir_ls(proj_path("inst", "templates"), type = "file"))
+  if (is_empty(templates)) {
+    return(character())
+  }
+
+  choice <- utils::menu(
+    choices = templates,
+    title = "Which template do you want to edit? (0 to exit)"
+  )
+
+  templates[choice]
 }
 
 #' Open configuration files